🐛 use crc32 rather than srand
srand is only meant to be used once per process and VoTD is not truely
random.  It's meant to be reproducable but arbitary.  Therefore we're
using the the same CRC based on the time at midnight on the selected day,
and using this as an offset into both book, chapter, and verse.

This removes the annoying problem that the first time VoTD is requested
in a given day, we get a different one.