🩹 fix short header processing; we will use */*
An exception 406 is not thrown any more.  ie. '*' or an empty Accept header,
or 'ab' or under three characters will just end up as */* and work now.

There is bound to be a situation where somebody passes Accept: ab in the future,
but this is not _intended_ to work and is clearly wrong, and an RFC violation...
so don't do this.