🏗️ id of verse now contains translation
Fairly large architectural change, I thought it might need to happen
this means you're not accidentally caching information from 'included'
about the the wrong verse, because things may be different between
translations.

There is a new getPath() method which returns what id used to return on a Verse, it uses id internally, but that's an opaque implementation-detail.