🏗️ remove signal handlers
they don't work under this FCGI model -
it looks like we're fine in that the child processes restart
if the master PID is sent a SIGHUP.  OK, it's not ideal but it
looks like it will have to do.  It won't clean up our cache.

Will it mess up our uptime?  It appears not, in my testing.
We might need to wait to see if messes up load rotation.