📜 new utility -> yaml2json
This utility will allow us to read the config file (YAML) as if it were
JSON via a pipe.  The point is to avoid installing 'yq' because the .deb
package does not appear to be available under Debian.  We can therefore
use 'jq' instead.  It's a kludge, but it will save us a big dependency
headache.