✨ new expiry per cookie-backend-type feature
This probably doesn't work exactly how it needs to, since the master
expiry value doesn't do anything.  We need to know if it has been
changed from the default and we can't do that with a HASH.  Hmm.

Might need some thought.  Or we only use the master value and scrap the
per-backend expiry?