➕ switch to Log4Perl for log rotation
I tried very hard to get SIPHUP to work but something causes all
children to exit, rather than merely re-open the logs.  Let's simply
move to a different way of rotating logs, by using Log4Perl itself,
and because compression is not natively supported, and _might_ slow
down handling queries, just rely on logrotate for that, out of band,
and only on logs which are certainly not being rotated at that time.